# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ML Model Monitoring API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ModelVersionStatus = typeof ModelVersionStatus[keyof typeof ModelVersionStatus];


export const ModelVersionStatus = {
  active: 'active',
  degraded: 'degraded',
  rolled_back: 'rolled_back',
  shadow: 'shadow',
} as const;

export interface ModelVersion {
  id: number;
  name: string;
  version: string;
  status: ModelVersionStatus;
  deployedAt: string;
  dailyUsers: number;
  /** @nullable */
  description?: string | null;
}

export interface ModelVersionInput {
  name: string;
  version: string;
  description?: string;
  dailyUsers?: number;
}

export type DriftMetricDriftSeverity = typeof DriftMetricDriftSeverity[keyof typeof DriftMetricDriftSeverity];


export const DriftMetricDriftSeverity = {
  stable: 'stable',
  warning: 'warning',
  critical: 'critical',
} as const;

export interface DriftMetric {
  id: number;
  modelId: number;
  modelName: string;
  timestamp: string;
  psiScore: number;
  ksStatistic: number;
  chiSquareStat: number;
  driftSeverity: DriftMetricDriftSeverity;
  featureCount?: number;
}

export type FeatureDriftScoreFeatureType = typeof FeatureDriftScoreFeatureType[keyof typeof FeatureDriftScoreFeatureType];


export const FeatureDriftScoreFeatureType = {
  continuous: 'continuous',
  categorical: 'categorical',
} as const;

export type FeatureDriftScoreSeverity = typeof FeatureDriftScoreSeverity[keyof typeof FeatureDriftScoreSeverity];


export const FeatureDriftScoreSeverity = {
  stable: 'stable',
  warning: 'warning',
  critical: 'critical',
} as const;

export interface FeatureDriftScore {
  featureName: string;
  featureType: FeatureDriftScoreFeatureType;
  psiScore: number;
  ksStatistic: number;
  severity: FeatureDriftScoreSeverity;
  nullRate: number;
  baselineNullRate: number;
}

export type DriftSummaryOverallSeverity = typeof DriftSummaryOverallSeverity[keyof typeof DriftSummaryOverallSeverity];


export const DriftSummaryOverallSeverity = {
  stable: 'stable',
  warning: 'warning',
  critical: 'critical',
} as const;

export interface DriftSummary {
  modelId: number;
  modelName: string;
  overallSeverity: DriftSummaryOverallSeverity;
  psiScore: number;
  ksStatistic: number;
  driftedFeatures: number;
  totalFeatures: number;
  detectedAt: string;
}

export interface LatencyMetric {
  id: number;
  modelId: number;
  modelName: string;
  timestamp: string;
  p50Ms: number;
  p95Ms: number;
  p99Ms: number;
  requestCount: number;
  errorRate?: number;
}

export type CurrentLatencyStatus = typeof CurrentLatencyStatus[keyof typeof CurrentLatencyStatus];


export const CurrentLatencyStatus = {
  normal: 'normal',
  warning: 'warning',
  critical: 'critical',
} as const;

export interface CurrentLatency {
  modelId: number;
  modelName: string;
  p50Ms: number;
  p95Ms: number;
  p99Ms: number;
  status: CurrentLatencyStatus;
}

export type AlertSeverity = typeof AlertSeverity[keyof typeof AlertSeverity];


export const AlertSeverity = {
  warning: 'warning',
  critical: 'critical',
} as const;

export type AlertType = typeof AlertType[keyof typeof AlertType];


export const AlertType = {
  drift: 'drift',
  latency: 'latency',
  null_rate: 'null_rate',
  schema_mismatch: 'schema_mismatch',
  prediction_skew: 'prediction_skew',
} as const;

export type AlertStatus = typeof AlertStatus[keyof typeof AlertStatus];


export const AlertStatus = {
  active: 'active',
  resolved: 'resolved',
} as const;

export interface Alert {
  id: number;
  modelId: number;
  modelName: string;
  severity: AlertSeverity;
  type: AlertType;
  message: string;
  triggeredAt: string;
  /** @nullable */
  resolvedAt?: string | null;
  status: AlertStatus;
  /** @nullable */
  runbookUrl?: string | null;
}

export type FeatureStatFeatureType = typeof FeatureStatFeatureType[keyof typeof FeatureStatFeatureType];


export const FeatureStatFeatureType = {
  continuous: 'continuous',
  categorical: 'categorical',
} as const;

export interface FeatureStat {
  id: number;
  modelId: number;
  featureName: string;
  featureType: FeatureStatFeatureType;
  nullRate: number;
  baselineNullRate: number;
  schemaMismatch: boolean;
  lastUpdated: string;
}

export type DashboardSummaryOverallHealth = typeof DashboardSummaryOverallHealth[keyof typeof DashboardSummaryOverallHealth];


export const DashboardSummaryOverallHealth = {
  healthy: 'healthy',
  degraded: 'degraded',
  critical: 'critical',
} as const;

export interface DashboardSummary {
  totalModels: number;
  activeModels: number;
  degradedModels: number;
  activeAlerts: number;
  criticalAlerts: number;
  mttdMinutes: number;
  totalDailyUsers: number;
  overallHealth: DashboardSummaryOverallHealth;
  recentDriftDetections?: number;
  preventedImpacts?: number;
}

export interface PredictionBin {
  bin: string;
  label: string;
  currentCount: number;
  baselineCount: number;
}

export type ListDriftMetricsParams = {
modelId?: number;
hours?: number;
};

export type GetFeatureHeatmapParams = {
modelId?: number;
};

export type GetDriftSummaryParams = {
modelId?: number;
};

export type ListLatencyMetricsParams = {
modelId?: number;
hours?: number;
};

export type ListAlertsParams = {
status?: ListAlertsStatus;
modelId?: number;
};

export type ListAlertsStatus = typeof ListAlertsStatus[keyof typeof ListAlertsStatus];


export const ListAlertsStatus = {
  active: 'active',
  resolved: 'resolved',
  all: 'all',
} as const;

export type ListFeaturesParams = {
modelId?: number;
};

export type GetPredictionDistributionParams = {
modelId?: number;
};