# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ML Model Monitoring API
 * OpenAPI spec version: 0.1.0
 */
import type { AlertSeverity } from './alertSeverity';
import type { AlertStatus } from './alertStatus';
import type { AlertType } from './alertType';

export interface Alert {
  id: number;
  modelId: number;
  modelName: string;
  severity: AlertSeverity;
  type: AlertType;
  message: string;
  triggeredAt: string;
  /** @nullable */
  resolvedAt?: string | null;
  status: AlertStatus;
  /** @nullable */
  runbookUrl?: string | null;
}