# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ML Model Monitoring API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryOverallHealth } from './dashboardSummaryOverallHealth';

export interface DashboardSummary {
  totalModels: number;
  activeModels: number;
  degradedModels: number;
  activeAlerts: number;
  criticalAlerts: number;
  mttdMinutes: number;
  totalDailyUsers: number;
  overallHealth: DashboardSummaryOverallHealth;
  recentDriftDetections?: number;
  preventedImpacts?: number;
}