# lib/api-zod/src/generated/types/driftMetric.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ML Model Monitoring API
 * OpenAPI spec version: 0.1.0
 */
import type { DriftMetricDriftSeverity } from './driftMetricDriftSeverity';

export interface DriftMetric {
  id: number;
  modelId: number;
  modelName: string;
  timestamp: string;
  psiScore: number;
  ksStatistic: number;
  chiSquareStat: number;
  driftSeverity: DriftMetricDriftSeverity;
  featureCount?: number;
}