# lib/api-zod/src/generated/types/driftSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ML Model Monitoring API
 * OpenAPI spec version: 0.1.0
 */
import type { DriftSummaryOverallSeverity } from './driftSummaryOverallSeverity';

export interface DriftSummary {
  modelId: number;
  modelName: string;
  overallSeverity: DriftSummaryOverallSeverity;
  psiScore: number;
  ksStatistic: number;
  driftedFeatures: number;
  totalFeatures: number;
  detectedAt: string;
}