# lib/api-zod/src/generated/types/featureStat.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ML Model Monitoring API
 * OpenAPI spec version: 0.1.0
 */
import type { FeatureStatFeatureType } from './featureStatFeatureType';

export interface FeatureStat {
  id: number;
  modelId: number;
  featureName: string;
  featureType: FeatureStatFeatureType;
  nullRate: number;
  baselineNullRate: number;
  schemaMismatch: boolean;
  lastUpdated: string;
}